src/requirements.txt
requests>=2.32.5
beautifulsoup4>=4.14.3
schedule>=1.2.2
python-dotenv>=1.2.1
pyTelegramBotAPI>=4.29.1